Update_20230224

+++ src/BookStore.Domain/Interfaces/ICategoryService.cs
@@ -10,7 +10,7 @@ namespace BookStore.Domain.Interfaces
     public interface ICategoryService
     {
         Task<IEnumerable<Category>> GetAll();
-        Task<Category> GetById(int id);
+        Task<Category> GetById(long id);
         Task<Category> Add(Category category);
         Task<Category> Update(Category category);
         Task<bool> Remove(Category category);
